refactor bin script
