+++ lib/evaluate.js
@@ -1,9 +1,12 @@
 import { ethers } from 'ethers'
-import { ieContractWithSigner } from './contract.js'
 
 const { BigNumber } = ethers
 
-export const evaluate = async ({ rounds, roundIndex }) => {
+export const evaluate = async ({
+  rounds,
+  roundIndex,
+  ieContractWithSigner
+}) => {
   // Get measurements
   const measurements = rounds[roundIndex] || []
