+++ lib/preprocess.js
@@ -1,16 +1,9 @@
-import { Web3Storage } from 'web3.storage'
 import assert from 'node:assert'
 import { ethers } from 'ethers'
 
-const { WEB3_STORAGE_API_TOKEN } = process.env
-
-assert(WEB3_STORAGE_API_TOKEN, 'WEB3_STORAGE_API_TOKEN required')
-
-const web3Storage = new Web3Storage({ token: WEB3_STORAGE_API_TOKEN })
-
 // Preprocess
-export const preprocess = async ({ rounds, cid, roundIndex }) => {
-  const measurements = (await fetchMeasurements(cid))
+export const preprocess = async ({ rounds, cid, roundIndex, web3Storage }) => {
+  const measurements = (await fetchMeasurements(web3Storage, cid))
     .filter(measurement => {
       try {
         assertValidMeasurement(measurement)
@@ -29,7 +22,7 @@ export const preprocess = async ({ rounds, cid, roundIndex }) => {
 }
 
 // Fetch measurements
-const fetchMeasurements = async cid => {
+const fetchMeasurements = async (web3Storage, cid) => {
   const res = await web3Storage.get(cid)
   const files = await res.files()
   const measurements = JSON.parse(await files[0].text())
